pyproject.toml
[tool.poetry]
name = "winds-mobi-admin"
description = "Django application to manage winds.mobi users and profiles."
version = "0.0.0"
authors = ["winds.mobi"]
license = "	AGPL-3.0-only"
packages = [
    { include = "winds_mobi_admin" },
    { include = "winds_mobi_user" },
    { include = "winds_mobi_zermatt" },
]

[tool.poetry.dependencies]
python = "3.9.*"

django = "3.2.13"
dj-database-url = "0.5.0"
djangorestframework = "3.13.1"
django-cors-headers = "3.13.0"
google-auth = "1.6.3"
google-auth-oauthlib = "0.3.0"
gunicorn = "20.1.0"
psycopg2 = "2.9.3"
pymongo = "3.12.2"
pyjwt = "2.4.0"
redis = "4.3.3"
sentry-sdk = "1.6.0"
whitenoise = "5.3.0"

[tool.poetry.dev-dependencies]
black = "22.3.0"
flake8 = "4.0.1"
python-dotenv = "^0.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
force-exclude  = '''
'''
